Remove osu web service bucket
