+++ app/api/dependencies.py
@@ -34,20 +34,20 @@ from app.repositories.tourney_pool_maps import TourneyPoolMapsRepository
 from app.repositories.tourney_pools import TourneyPoolsRepository
 from app.repositories.user_achievements import UserAchievementsRepository
 from app.repositories.users import UsersRepository
+from app.services.accounts import AccountRegistrationService
 from app.services.bancho import BanchoLoginService
 from app.services.beatmap_leaderboards import BeatmapLeaderboardService
 from app.services.clans import ClansService
 from app.services.client_integrity import ClientIntegrityService
+from app.services.comments import CommentsService
 from app.services.direct_search import DirectSearchParams
 from app.services.direct_search import DirectSearchService
+from app.services.favourites import FavouritesService
+from app.services.mail import MailReadService
+from app.services.maps import BeatmapInfoService
+from app.services.maps import BeatmapRatingService
+from app.services.maps import BeatmapSetService
 from app.services.maps import MapsService
-from app.services.osu_web import AccountRegistrationService
-from app.services.osu_web import BeatmapInfoService
-from app.services.osu_web import BeatmapRatingService
-from app.services.osu_web import BeatmapSetService
-from app.services.osu_web import CommentsService
-from app.services.osu_web import FavouritesService
-from app.services.osu_web import MailReadService
 from app.services.performance import PerformanceService
 from app.services.players import PlayersService
 from app.services.public_api import PublicApiService
