+++ app/api/domains/osu.py
@@ -43,25 +43,25 @@ from app.objects.player import ModeData
 from app.objects.player import Player
 from app.objects.score import Score
 from app.repositories.achievements import Achievement
+from app.services.accounts import AccountRegistrationResultCode
+from app.services.accounts import AccountRegistrationService
 from app.services.beatmap_leaderboards import BeatmapLeaderboardRequest
 from app.services.beatmap_leaderboards import BeatmapLeaderboardResult
 from app.services.beatmap_leaderboards import BeatmapLeaderboardResultCode
 from app.services.beatmap_leaderboards import BeatmapLeaderboardService
 from app.services.client_integrity import ClientIntegrityResult
 from app.services.client_integrity import ClientIntegrityService
+from app.services.comments import CommentsService
 from app.services.direct_search import DirectSearchResult
 from app.services.direct_search import DirectSearchResultCode
 from app.services.direct_search import DirectSearchService
-from app.services.osu_web import AccountRegistrationResultCode
-from app.services.osu_web import AccountRegistrationService
-from app.services.osu_web import AddFavouriteResult
-from app.services.osu_web import BeatmapInfoService
-from app.services.osu_web import BeatmapRatingResultCode
-from app.services.osu_web import BeatmapRatingService
-from app.services.osu_web import BeatmapSetService
-from app.services.osu_web import CommentsService
-from app.services.osu_web import FavouritesService
-from app.services.osu_web import MailReadService
+from app.services.favourites import AddFavouriteResult
+from app.services.favourites import FavouritesService
+from app.services.mail import MailReadService
+from app.services.maps import BeatmapInfoService
+from app.services.maps import BeatmapRatingResultCode
+from app.services.maps import BeatmapRatingService
+from app.services.maps import BeatmapSetService
 from app.services.replays import ReplayResultCode
 from app.services.replays import ReplayService
 from app.services.score_submission import ScoreSubmissionError
